# internal/core/entities/product.go
package entities

import (
	"time"

	"go.mongodb.org/mongo-driver/bson/primitive"
)

// Product represents a product entity in the system
type Product struct {
	ID        primitive.ObjectID `bson:"_id,omitempty" json:"id"`
	Name      string             `bson:"name" json:"name"`
	Price     float32            `bson:"price" json:"price"`
	CreatedAt time.Time          `bson:"created_at" json:"created_at"`
	UpdatedAt time.Time          `bson:"updated_at" json:"updated_at"`
}